Removed redundant verifications in services and commands.

+++ src/main/java/com/jayton/admissionoffice/command/impl/user/UserCancelApplicationCommand.java
@@ -4,6 +4,7 @@ import com.jayton.admissionoffice.command.Command;
 import com.jayton.admissionoffice.command.exception.ShownException;
 import com.jayton.admissionoffice.command.exception.VerificationException;
 import com.jayton.admissionoffice.command.util.Verifier;
+import com.jayton.admissionoffice.model.to.Application;
 import com.jayton.admissionoffice.model.to.SessionTerms;
 import com.jayton.admissionoffice.model.user.User;
 import com.jayton.admissionoffice.service.ApplicationService;
@@ -35,19 +36,13 @@ public class UserCancelApplicationCommand implements Command {
             HttpSession session = request.getSession();
             User logged = (User) session.getAttribute(PARAM_NAMES.getString("user"));
 
-            LocalDateTime current = LocalDateTime.now();
-
-            SessionTerms sessionTerms = utilService.getSessionTerms((short)current.getYear());
-            if(current.isAfter(sessionTerms.getSessionEnd()) || current.isBefore(sessionTerms.getSessionStart())) {
-                logger.error("Failed to cancel application.");
-                request.setAttribute(PARAM_NAMES.getString("shownException"),
-                        new ShownException("Can not cancel application after session term expiration."));
-                response.sendRedirect(PAGE_NAMES.getString("controller.get_user")+"&id="+logged.getId());
-            }
-
             Long id = Verifier.convertToLong(request.getParameter(PARAM_NAMES.getString("id")));
             Verifier.verifyId(id);
 
+            Application removed = applicationService.get(id);
+            if(removed.getUserId() != logged.getId()) {
+                throw new VerificationException("Access denied");
+            }
             applicationService.delete(id);
 
             response.sendRedirect(PAGE_NAMES.getString("controller.get_user")+"&id="+logged.getId());
